requirements.txt
# ============================================================
# TEKNOFEST 2026 — Onkolojide 3T (GlioSight)
# Python Bağımlılıkları
# ============================================================

# --- Derin Öğrenme ---
torch>=2.1.0
torchvision>=0.16.0
monai>=1.3.0          # Medical image analysis (3D U-Net, nnU-Net)

# --- Görüntü İşleme & Tıbbi Veri ---
nibabel>=5.1.0        # NIfTI formatı (MRI dosyaları)
SimpleITK>=2.3.0      # Medikal görüntü işleme
pydicom>=2.4.0        # DICOM okuma
scikit-image>=0.21.0

# --- Veri Bilimi ---
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# --- Survival Analysis ---
lifelines>=0.27.0     # Kaplan-Meier, Cox PH modeli
scikit-survival>=0.21.0

# --- Gradient Boosting ---
xgboost>=2.0.0
lightgbm>=4.1.0

# --- Açıklanabilir AI ---
shap>=0.43.0
grad-cam>=1.4.8       # GradCAM, GradCAM++ vb.

# --- Radyomik ---
pyradiomics>=3.1.0    # Radyomik özellik çıkarımı

# --- Görselleştirme ---
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
nilearn>=0.10.0       # Beyin görüntüleme görselleştirme

# --- Konfigürasyon ---
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0.0

# --- Loglama & Deney Takibi ---
mlflow>=2.8.0
tensorboard>=2.14.0
tqdm>=4.66.0
rich>=13.6.0

# --- Notebook ---
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# --- Test ---
pytest>=7.4.0
pytest-cov>=4.1.0

# --- Kodlama Kalitesi ---
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
